Паттерн [+] Переписал экран с добавление или редактированием заметки с использованием паттерна MVP.
