+++ app/src/main/java/ru/plumsoftware/notebook/presentation/activities/note/model/Mode.java
@@ -1,0 +1,6 @@
+package ru.plumsoftware.notebook.presentation.activities.note.model;
+
+public enum Mode {
+    New,
+    Edit
+}
