+++ app/src/main/java/ru/plumsoftware/notebook/presentation/activities/note/view/AddNoteView.java
@@ -1,0 +1,9 @@
+package ru.plumsoftware.notebook.presentation.activities.note.view;
+
+import ru.plumsoftware.data.model.ui.Note;
+
+public interface AddNoteView {
+    void initNote(Note note);
+    void initToolbarTitle(String title, String textOnButton, String time);
+    void showSnackBar();
+}
