+++ app/src/main/res/layout/activity_add_note.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".presentation.activities.AddNoteActivity">
+    tools:context=".presentation.activities.note.view.AddNoteActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
