+++ app/src/main/res/layout/activity_main.xml
@@ -12,10 +12,11 @@
         android:id="@+id/addNote"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_margin="16dp"
+        android:layout_margin="18dp"
         android:layout_marginEnd="8dp"
         android:layout_marginBottom="8dp"
-        android:background="@drawable/button_click_3"
+        android:background="@drawable/button_click_2"
+        android:contentDescription=""
         android:elevation="4dp"
         android:padding="10dp"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -24,7 +25,8 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="1.0"
-        app:srcCompat="@drawable/ic_add" />
+        app:srcCompat="@drawable/ic_add"
+        tools:ignore="ContentDescription" />
 
     <LinearLayout
         android:id="@+id/linearLayout"
