+++ app/src/main/res/xml-v25/shortcuts.xml
@@ -10,7 +10,7 @@
         <intent
             android:action="android.intent.action.VIEW"
             android:targetPackage="ru.plumsoftware.notebook"
-            android:targetClass="ru.plumsoftware.notebook.presentation.activities.AddNoteActivity" />
+            android:targetClass="ru.plumsoftware.notebook.presentation.activities.note.view.AddNoteActivity" />
         <categories android:name="android.shortcut.conversation" />
         <capability-binding android:key="actions.intent.CREATE_MESSAGE" />
     </shortcut>
